+++ src/node/commit.rs
@@ -264,3 +264,225 @@ impl<J: Jet> CommitNode<J> {
         program_and_witness_bytes
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    use bitcoin_hashes::hex::ToHex;
+    use std::fmt;
+
+    use crate::decode::Error;
+    use crate::jet::Core;
+    use crate::node::SimpleFinalizer;
+    use crate::{BitMachine, Value};
+
+    fn assert_program_deserializable<J: Jet>(
+        prog_bytes: &[u8],
+        cmr_str: &str,
+    ) -> Arc<CommitNode<J>> {
+        let prog_hex = prog_bytes.to_hex();
+
+        let mut iter = BitIter::from(prog_bytes);
+        let prog = match CommitNode::<J>::decode(&mut iter) {
+            Ok(prog) => prog,
+            Err(e) => panic!("program {} failed: {}", prog_hex, e),
+        };
+
+        assert_eq!(
+            prog.cmr().to_string(),
+            cmr_str,
+            "CMR mismatch (got {} expected {}) for program {}",
+            prog.cmr().to_string(),
+            cmr_str,
+            prog_hex,
+        );
+
+        let reser_sink = prog.encode_to_vec();
+        assert_eq!(
+            prog_bytes,
+            &reser_sink[..],
+            "program {} reserialized as {}",
+            prog_hex,
+            reser_sink.to_hex(),
+        );
+
+        prog
+    }
+
+    fn assert_program_not_deserializable<J: Jet>(prog: &[u8], err: &dyn fmt::Display) {
+        let prog_hex = prog.to_hex();
+        let err_str = err.to_string();
+
+        let mut iter = BitIter::from(prog);
+        match CommitNode::<J>::decode(&mut iter) {
+            Ok(prog) => panic!(
+                "Program {} succeded (expected error {}). Program parsed as:\n{}",
+                prog_hex, err, prog
+            ),
+            Err(e) if e.to_string() == err_str => {} // ok
+            Err(e) => panic!(
+                "Program {} failed with error {} (expected error {})",
+                prog_hex, e, err
+            ),
+        };
+    }
+
+    #[test]
+    fn canonical_order() {
+        // "main = comp unit iden", but with the iden serialized before the unit
+        // To obtain this test vector I temporarily swapped `get_left` and `get_right`
+        // in the implementation of `PostOrderIter`
+        assert_program_not_deserializable::<Core>(&[0xa8, 0x48, 0x10], &Error::NotInCanonicalOrder);
+
+        // "main = iden", but prefixed by some unused nodes, the first of which is also iden.
+        assert_program_not_deserializable::<Core>(
+            &[0xc1, 0x00, 0x06, 0x20],
+            &Error::NotInCanonicalOrder,
+        );
+    }
+
+    #[test]
+    fn hidden_node() {
+        // main = hidden deadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeef
+        #[rustfmt::skip]
+        let hidden = [
+            0x36, 0xf5, 0x6d, 0xf7, 0x7e, 0xf5, 0x6d, 0xf7,
+            0x7e, 0xf5, 0x6d, 0xf7, 0x7e, 0xf5, 0x6d, 0xf7,
+            0x7e, 0xf5, 0x6d, 0xf7, 0x7e, 0xf5, 0x6d, 0xf7,
+            0x7e, 0xf5, 0x6d, 0xf7, 0x7e, 0xf5, 0x6d, 0xf7,
+            78,
+        ];
+        assert_program_not_deserializable::<Core>(&hidden, &Error::HiddenNode);
+
+        // main = comp witness hidden deadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeef
+        let hidden = [
+            0xae, 0xdb, 0xd5, 0xb7, 0xdd, 0xfb, 0xd5, 0xb7, 0xdd, 0xfb, 0xd5, 0xb7, 0xdd, 0xfb,
+            0xd5, 0xb7, 0xdd, 0xfb, 0xd5, 0xb7, 0xdd, 0xfb, 0xd5, 0xb7, 0xdd, 0xfb, 0xd5, 0xb7,
+            0xdd, 0xfb, 0xd5, 0xb7, 0xdd, 0xe0, 0x80,
+        ];
+        assert_program_not_deserializable::<Core>(&hidden, &Error::HiddenNode);
+    }
+
+    #[test]
+    fn case_both_children_hidden() {
+        // h1 = hidden deadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeef
+        // main = case h1 h1
+        #[rustfmt::skip]
+        let hidden = [
+            0x8d, 0xbd, 0x5b, 0x7d, 0xdf, 0xbd, 0x5b, 0x7d,
+            0xdf, 0xbd, 0x5b, 0x7d, 0xdf, 0xbd, 0x5b, 0x7d,
+            0xdf, 0xbd, 0x5b, 0x7d, 0xdf, 0xbd, 0x5b, 0x7d,
+            0xdf, 0xbd, 0x5b, 0x7d, 0xdf, 0xbd, 0x5b, 0x7d,
+            0xde, 0x10,
+        ];
+        assert_program_not_deserializable::<Core>(&hidden, &Error::BothChildrenHidden);
+    }
+
+    #[test]
+    fn unshared_hidden() {
+        // This program has a repeated hidden node, but all other sharing is correct
+        // and the order is canonical, etc.
+        #[rustfmt::skip]
+        let hidden = [
+            0xd6, 0xe9, 0x62, 0x56, 0x62, 0xc9, 0x38, 0x8a,
+            0x44, 0x31, 0x85, 0xee, 0xc2, 0x2b, 0x91, 0x48,
+            0x87, 0xe1, 0xfd, 0x18, 0x57, 0xc2, 0x8c, 0x4a,
+            0x28, 0x44, 0x2f, 0xa8, 0x61, 0x5c, 0xa7, 0x6e,
+            0x8c, 0xf9, 0x80, 0xc2, 0x18, 0x95, 0x98, 0xb2,
+            0x4e, 0x22, 0x91, 0x0c, 0x61, 0x7b, 0xb0, 0x8a,
+            0xe4, 0x52, 0x21, 0xf8, 0x7f, 0x46, 0x15, 0xf0,
+            0xa3, 0x12, 0x8a, 0x11, 0x0b, 0xea, 0x18, 0x57,
+            0x29, 0xdb, 0xa3, 0x3e, 0x60, 0x30, 0x2c, 0x00,
+            0xd0, 0x48, 0x20,
+        ];
+        assert_program_not_deserializable::<Core>(&hidden, &Error::SharingNotMaximal);
+    }
+
+    #[test]
+    fn shared_witnesses() {
+        // main = witness :: A -> B
+        assert_program_deserializable::<Core>(
+            &[0x38],
+            "bf12681a76fc7c00c63e583c25cc97237337d6aca30d3f4a664075445385c648",
+        );
+
+        // # Program which demands two 32-bit witnesses, the first one == the second + 1
+        // wit1 = witness :: 1 -> 2^32
+        // wit2 = witness :: 1 -> 2^32
+        //
+        // wit_diff = comp (comp (pair wit1 wit2) jet_subtract_32) (drop iden) :: 1 -> 2^32
+        // diff_is_one = comp (pair wit_diff jet_one_32) jet_eq_32             :: 1 -> 2
+        // main = comp diff_is_one jet_verify                                  :: 1 -> 1
+
+        #[rustfmt::skip]
+        let bad_diff1s = vec![
+            // Above program, but with both witness nodes shared (note they have
+            // the same type and CMR)
+            vec![
+                0xda, 0xe2, 0x39, 0xa3, 0x10, 0x42, 0x0e, 0x05,
+                0x71, 0x88, 0xa3, 0x6d, 0xc4, 0x11, 0x80, 0x80
+            ],
+            // Same program but with each `witness` replaced by `comp iden witness`, which
+            // is semantically the same but buries the offending witness nodes a bit to
+            // trip up naive sharing logic.
+            vec![
+                0xde, 0x87, 0x04, 0x08, 0xe6, 0x8c, 0x41, 0x08,
+                0x38, 0x15, 0xc6, 0x22, 0x8d, 0xb7, 0x10, 0x46,
+                0x02, 0x00,
+            ],
+        ];
+        for bad_diff1 in bad_diff1s {
+            assert_program_not_deserializable::<Core>(&bad_diff1, &Error::SharingNotMaximal);
+        }
+
+        #[rustfmt::skip]
+        let diff1s = vec![
+            (
+                // Sharing corrected
+                vec![
+                    0xdc, 0xee, 0x28, 0xe6, 0x8c, 0x41, 0x08, 0x38,
+                    0x15, 0xc6, 0x22, 0x8d, 0xb7, 0x10, 0x46, 0x02,
+                    0x00,
+                ],
+                // CMR not checked against C code, since C won't give us any data without witnesses
+                "a2ad9852818c0dc9307b476464cb9366c5c97896ba128f2f526b51910218293c",
+            ),
+            // Same program but with each `witness` replaced by `comp iden witness`.
+            (
+                vec![
+                    0xe0, 0x28, 0x70, 0x43, 0x83, 0x00, 0xab, 0x9a,
+                    0x31, 0x04, 0x20, 0xe0, 0x57, 0x18, 0x8a, 0x36,
+                    0xdc, 0x41, 0x18, 0x08, 
+                ],
+                // CMR not checked against C code, since C won't give us any data without witnesses
+                "f4583eca2a35aa48e8895235b58cfe90ba2196fbf7722b7d847c3c55eb6bdc0e",
+            )
+        ];
+
+        for (diff1, cmr) in diff1s {
+            let diff1_prog =
+                crate::node::commit::tests::assert_program_deserializable::<Core>(&diff1, cmr);
+
+            // Attempt to finalize, providing 32-bit witnesses 0, 1, ..., and then
+            // counting how many were consumed afterward.
+            let mut counter = 0..100;
+            let witness_iter = (&mut counter).rev().map(Value::u32).map(Arc::new);
+            let diff1_final = diff1_prog
+                .finalize(&mut SimpleFinalizer::new(witness_iter))
+                .unwrap();
+            assert_eq!(counter, 0..98);
+
+            // Execute the program to confirm that it worked
+            let mut mac = BitMachine::for_program(&diff1_final);
+            mac.exec(&diff1_final, &()).unwrap();
+        }
+    }
+
+    #[test]
+    fn extra_nodes() {
+        // main = comp unit unit # but with an extra unconnected `unit` stuck on the beginning
+        // I created this unit test by hand
+        assert_program_not_deserializable::<Core>(&[0xa9, 0x48, 0x00], &Error::NotInCanonicalOrder);
+    }
+}
