+++ src/node/redeem.rs
@@ -315,8 +315,82 @@ impl<J: Jet> RedeemNode<J> {
 #[cfg(test)]
 mod tests {
     use super::*;
+
+    use bitcoin_hashes::hex::ToHex;
+    use std::fmt;
+
     use crate::jet::Core;
     use crate::node::SimpleFinalizer;
+
+    fn assert_program_deserializable<J: Jet>(
+        prog_bytes: &[u8],
+        cmr_str: &str,
+        amr_str: &str,
+        imr_str: &str,
+    ) -> Arc<RedeemNode<J>> {
+        let prog_hex = prog_bytes.to_hex();
+
+        let mut iter = BitIter::from(prog_bytes);
+        let prog = match RedeemNode::<J>::decode(&mut iter) {
+            Ok(prog) => prog,
+            Err(e) => panic!("program {} failed: {}", prog_hex, e),
+        };
+
+        assert_eq!(
+            prog.cmr().to_string(),
+            cmr_str,
+            "CMR mismatch (got {} expected {}) for program {}",
+            prog.cmr().to_string(),
+            cmr_str,
+            prog_hex,
+        );
+
+        assert_eq!(
+            prog.amr().to_string(),
+            amr_str,
+            "AMR mismatch (got {} expected {}) for program {}",
+            prog.amr().to_string(),
+            amr_str,
+            prog_hex,
+        );
+        assert_eq!(
+            prog.imr().to_string(),
+            imr_str,
+            "IMR mismatch (got {} expected {}) for program {}",
+            prog.imr().to_string(),
+            imr_str,
+            prog_hex,
+        );
+
+        let reser_sink = prog.encode_to_vec();
+        assert_eq!(
+            prog_bytes,
+            &reser_sink[..],
+            "program {} reserialized as {}",
+            prog_hex,
+            reser_sink.to_hex(),
+        );
+
+        prog
+    }
+
+    fn assert_program_not_deserializable<J: Jet>(prog: &[u8], err: &dyn fmt::Display) {
+        let prog_hex = prog.to_hex();
+        let err_str = err.to_string();
+
+        let mut iter = BitIter::from(prog);
+        match RedeemNode::<J>::decode(&mut iter) {
+            Ok(prog) => panic!(
+                "Program {} succeded (expected error {}). Program parsed as:\n{}",
+                prog_hex, err, prog
+            ),
+            Err(e) if e.to_string() == err_str => {} // ok
+            Err(e) => panic!(
+                "Program {} failed with error {} (expected error {})",
+                prog_hex, e, err
+            ),
+        };
+    }
 
     #[test]
     fn encode_shared_witnesses() {
@@ -348,11 +422,14 @@ mod tests {
         // This program is exactly the output from the `encode_shared_witnesses` test.
         // The point of this is to make sure that our witness-unsharing logic doesn't
         // get confused here and try to read two witnesses when there are only one.
-        let eqwits = vec![
-            0xc9, 0xc4, 0x6d, 0xb8, 0x82, 0x30, 0x11, 0xe2, 0x0d, 0xea, 0xdb, 0xee, 0xf0,
-        ];
-        let mut iter = BitIter::from(&eqwits[..]);
-        RedeemNode::<crate::jet::Core>::decode(&mut iter).unwrap();
+        assert_program_deserializable::<Core>(
+            &[
+                0xc9, 0xc4, 0x6d, 0xb8, 0x82, 0x30, 0x11, 0xe2, 0x0d, 0xea, 0xdb, 0xee, 0xf0,
+            ],
+            "e552742731de7f5c3c83cd54176c0ca6acf6dbd3c37bef7da132eb06f3856d06",
+            "6522c14daf762467232b55c37df25c17f80d7c12000ce8a5dcbbd67b879d192f",
+            "bbb31a66836ab67de2ef0efa1ce82078b36c9dbcf6af77822b87f7743dc88a8f",
+        );
     }
 
     #[test]
@@ -362,13 +439,10 @@ mod tests {
         // id2 = iden          :: A -> A # cmr dbfefcfc...
         // cp3 = comp id1 id2  :: A -> A # cmr c1ae55b5...
         // main = comp cp3 cp3 :: A -> A # cmr 314e2879...
-        let bad = [0xc1, 0x08, 0x04, 0x00, 0x00, 0x74, 0x74, 0x74];
-        let mut iter = BitIter::from(&bad[..]);
-        let err = RedeemNode::<crate::jet::Core>::decode(&mut iter).unwrap_err();
-        assert!(matches!(
-            err,
-            Error::Decode(crate::decode::Error::SharingNotMaximal)
-        ));
+        assert_program_not_deserializable::<Core>(
+            &[0xc1, 0x08, 0x04, 0x00, 0x00, 0x74, 0x74, 0x74],
+            &Error::Decode(crate::decode::Error::SharingNotMaximal),
+        );
     }
 
     #[test]
@@ -383,5 +457,166 @@ mod tests {
         } else {
             panic!("accepted program with bad witness length")
         }
+    }
+
+    #[test]
+    fn shared_grandchild() {
+        // # This program repeats the node `cp2` three times; during iteration it will
+        // # be placed on the stack as part of the initial `comp` combinator, but by
+        // # the time we get to it, it will have already been yielded. Makes sure this
+        // # does not confuse the iteration logic and break the decoded program structure.
+        // id1 = iden
+        // cp2 = comp id1 id1
+        // cp3 = comp cp2 cp2
+        // main = comp cp3 cp2
+        assert_program_deserializable::<Core>(
+            &[0xc1, 0x00, 0x00, 0x01, 0x00],
+            "c2c86be0081a9c75af49098f359c7efdfa7ccbd0459adb11bcf676b80c8644b1",
+            "e053520f0c3219d1cabd705b4523ccd05c8d703a70f6f3994a20774a42b5ccfc",
+            "7b0ad0514279280d5c2ac1da729222936b8768d9f465c6c6ade3b0ed7dc97263",
+        );
+    }
+
+    #[test]
+    #[rustfmt::skip]
+    fn assert_lr() {
+        // asst = assertl unit deadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeef
+        // input0 = pair (injl unit) unit
+        // main = comp input0 asst
+        assert_program_deserializable::<Core>(
+            &[
+                0xcd, 0x24, 0x08, 0x4b, 0x6f, 0x56, 0xdf, 0x77,
+                0xef, 0x56, 0xdf, 0x77, 0xef, 0x56, 0xdf, 0x77,
+                0xef, 0x56, 0xdf, 0x77, 0xef, 0x56, 0xdf, 0x77,
+                0xef, 0x56, 0xdf, 0x77, 0xef, 0x56, 0xdf, 0x77,
+                0xef, 0x56, 0xdf, 0x77, 0x86, 0x01, 0x80,
+            ],
+            "c7194362a5480900dd44f9f647a49b8adcb92a25fb293c920e6bbcf6977cf63d",
+            "eaf95c23d967563132b65e43578fe08dae2a29ac66775ddd37af3ac7de28678b",
+            "d2927a9a54ddea8359ee00aa27e0aa1e354cc6924b090c759e2ed686712700a0",
+        );
+
+
+        // asst = assertr deadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeefdeadbeef unit
+        // input1 = pair (injr unit) unit
+        // main = comp input1 asst
+        assert_program_deserializable::<Core>(
+            &[
+                0xcd, 0x25, 0x08, 0x6d, 0xea, 0xdb, 0xee, 0xfd,
+                0xea, 0xdb, 0xee, 0xfd, 0xea, 0xdb, 0xee, 0xfd,
+                0xea, 0xdb, 0xee, 0xfd, 0xea, 0xdb, 0xee, 0xfd,
+                0xea, 0xdb, 0xee, 0xfd, 0xea, 0xdb, 0xee, 0xfd,
+                0xea, 0xdb, 0xee, 0xf4, 0x86, 0x01, 0x80,
+            ],
+            "8e471ac519e0b16a2b7dda7e8d68165f260cae4823861ddc494b7c73a615b212",
+            "ea1ee417816a57b80739520c7319c33a39a5f4ce7b59856e69f768d5d8f174a6",
+            "f262f83f1c9341390e015e4c5126f3954e17a1f275af73da2948eaf4797fda48",
+        );
+    }
+
+    #[test]
+    #[rustfmt::skip]
+    fn disconnect() {
+        // id1 = iden                 :: (2^256 * B) -> (2^256 * B)                       # cmr dbfefcfc...
+        // pr2 = pair id1 id1         :: (2^256 * B) -> ((2^256 * B) * (2^256 * B))       # cmr a62c628c...
+        // disc3 = disconnect pr2 pr2 :: B -> ((2^256 * B) * ((2^256 * B) * (2^256 * B))) # cmr d81d6f28...
+        // ut4 = unit                 :: ((2^256 * B) * ((2^256 * B) * (2^256 * B))) -> 1 # cmr 62274a89...
+        // main = comp disc3 ut4      :: B -> 1                                           # cmr a453360c...
+        assert_program_deserializable::<Core>(
+            &[0xc5, 0x02, 0x06, 0x24, 0x10],
+            "a453360c0825cc2d3c4c907d67b174273b0e0386c7e5ecdb28394a8f37fd68b9",
+            "d5b05a5da87ee490312279496e12e17bc987c98219d8961bc3a7c3ec95a7ce1e",
+            "3579ae2a05bbe689f16bd3ff29d840ae8aa8bbad70f6de27b7473746637abeb6",
+        );
+    }
+
+    #[test]
+    #[rustfmt::skip]
+    #[cfg(feature = "elements")]
+    fn disconnect2() {
+        // Program that Russell posted on IRC 2023-06-22 that tickled `Dag::right_child`
+        // bug that had been introduced that day. Probably not the most minimal test
+        // vector but might as well include it as it seems like an interesting program.
+        //
+        // # Witnesses
+        // wit1 = witness :: 1 -> 2^512
+        //
+        // # Constants
+        // const1 = word_jet 0x00000000000000000000003b78ce563f89a0ed9414f5aa28ad0d96d6795f9c63 :: 1 -> 2^256 # cmr a9e3dbca...
+        //
+        // # Program code
+        // id1 = iden                 :: (2^256 * 1) -> (2^256 * 1)     # cmr dbfefcfc...
+        // jt2 = jet_sig_all_hash     :: 1 -> 2^256                     # cmr 9902bc0f...
+        // disc3 = disconnect id1 jt2 :: 1 -> 2^512                     # cmr 6968f10e...
+        // pr4 = pair const1 disc3    :: 1 -> (2^256 * 2^512)           # cmr 378ad609...
+        // pr5 = pair pr4 wit1        :: 1 -> ((2^256 * 2^512) * 2^512) # cmr 0d51ff00...
+        // jt6 = jet_check_sig_verify :: ((2^256 * 2^512) * 2^512) -> 1 # cmr 297459d8...
+        //
+        // main = comp pr5 jt6        :: 1 -> 1                         # cmr 14a5e0cc...
+        assert_program_deserializable::<crate::jet::Elements>(
+            &[
+                0xd3, 0x69, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x3b, 0x78, 0xce,
+                0x56, 0x3f, 0x89, 0xa0, 0xed, 0x94, 0x14, 0xf5,
+                0xaa, 0x28, 0xad, 0x0d, 0x96, 0xd6, 0x79, 0x5f,
+                0x9c, 0x63, 0x47, 0x07, 0x02, 0xc0, 0xe2, 0x8d,
+                0x88, 0x11, 0xe9, 0x00, 0x00, 0x00, 0x00, 0x00,
+                0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x1d,
+                0xbc, 0x67, 0x2b, 0x1f, 0xc4, 0xd0, 0x76, 0xca,
+                0x0a, 0x7a, 0xd5, 0x14, 0x56, 0x86, 0xcb, 0x6b,
+                0x3c, 0xaf, 0xce, 0x31, 0xed, 0xc3, 0x46, 0xa2,
+                0xd0, 0x5e, 0x0e, 0x8c, 0x80, 0x98, 0x15, 0xe4,
+                0x3d, 0x43, 0x8e, 0x78, 0xac, 0x71, 0x5e, 0xf1,
+                0x67, 0xd3, 0x22, 0xd4, 0x4a, 0xe0, 0xda, 0x2e,
+                0xb4, 0x75, 0x12, 0x60, 0x00,
+            ],
+            "14a5e0cc13da9acdd5f758ae7186802137143e06c8dcba10019ffec790359ee7",
+            "1a2a1cb30027f64bd9a639bf385834f415cac889ea654b0e04e81b210d3502d4",
+            "6de2a91db704fa287d8724da2e43a087bbd91334b0fb7160b231129356dc03b8",
+        );
+    }
+
+    #[test]
+    #[rustfmt::skip]
+    #[cfg(feature = "elements")]
+    fn disconnect3() {
+        // Yet another disconnect-based program that hit a bug in our AMR computation
+        // (passing left arrow in place of right arrow to the AMR constructor.)
+        // # Program code
+        // id1 = iden                 :: (2^256 * 1) -> (2^256 * 1) # cmr dbfefcfc...
+        // ut2 = unit                 :: 1 -> 1                     # cmr 62274a89...
+        // jl3 = injl ut2             :: 1 -> 2                     # cmr bd0cce93...
+        // disc4 = disconnect id1 jl3 :: 1 -> (2^256 * 2)           # cmr 6968f10e...
+        // ut5 = unit                 :: (2^256 * 2) -> 1           # cmr 62274a89...
+        // main = comp disc4 ut5      :: 1 -> 1                     # cmr a8c9cc7a...
+        assert_program_deserializable::<crate::jet::Elements>(
+            &[0xc9, 0x09, 0x20, 0x74, 0x90, 0x40],
+            "a8c9cc7a83518d0886afe1078d88eabca8353509e8c2e3b5c72cf559c713c9f5",
+            "97f77a7e7d7f3b2b1ac790bf54b39d47d6db8dcab7ed3c0a48df12f2c940af58",
+            "ed8152948589d65e0dea6d84f90eb752f63df818041f46bdc8f959f33299cbd3",
+        );
+    }
+
+    #[test]
+    #[cfg(feature = "elements")]
+    fn decode_schnorr() {
+        #[rustfmt::skip]
+        let schnorr0 = vec![
+            0xc6, 0xd5, 0xf2, 0x61, 0x14, 0x03, 0x24, 0xb1, 0x86, 0x20, 0x92, 0x68, 0x9f, 0x0b, 0xf1, 0x3a,
+            0xa4, 0x53, 0x6a, 0x63, 0x90, 0x8b, 0x06, 0xdf, 0x33, 0x61, 0x0c, 0x03, 0xe2, 0x27, 0x79, 0xc0,
+            0x6d, 0xf2, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+            0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00, 0x00,
+            0x00, 0x00, 0xe2, 0x8d, 0x8c, 0x04, 0x7a, 0x40, 0x1d, 0x20, 0xf0, 0x63, 0xf0, 0x10, 0x91, 0xa2,
+            0x0d, 0x34, 0xa6, 0xe3, 0x68, 0x04, 0x82, 0x06, 0xc9, 0x7b, 0xe3, 0x8b, 0xf0, 0x60, 0xf6, 0x01,
+            0x09, 0x8a, 0xbe, 0x39, 0xc5, 0xb9, 0x50, 0x42, 0xa4, 0xbe, 0xcd, 0x49, 0x50, 0xbd, 0x51, 0x6e,
+            0x3c, 0x90, 0x54, 0xe9, 0xe7, 0x05, 0xa5, 0x9c, 0xbd, 0x7d, 0xdd, 0x1f, 0xb6, 0x42, 0xe5, 0xe8,
+            0xef, 0xbe, 0x92, 0x01, 0xa6, 0x20, 0xa6, 0xd8, 0x00
+        ];
+        assert_program_deserializable::<crate::jet::Elements>(
+            &schnorr0,
+            "7bc56cb16d84999b977b58e1bc71dbe9edcc33650afc8a6ee05cfef8d608132b",
+            "02796d1d7d906a15d0a1ebed9d702e334b21e9cc52a578c24fca0fb7be82fac0",
+            "00ee3d3e7b7a65fcd77b6309c4d8464f176f13c92d8d69238eb0158f70c8a4df",
+        );
     }
 }
